+++ datenow/src/main/java/com/grepp/datenow/infra/error/exception/member/SessionExpiredException.java
@@ -1,0 +1,7 @@
+package com.grepp.datenow.infra.error.exception.member;
+
+public class SessionExpiredException extends RuntimeException {
+    public SessionExpiredException(String message) {
+        super(message);
+    }
+}
