+++ datenow/src/main/java/com/grepp/datenow/infra/error/exception/member/TokenMismatchException.java
@@ -1,0 +1,9 @@
+package com.grepp.datenow.infra.error.exception.member;
+
+public class TokenMismatchException extends RuntimeException {
+
+    public TokenMismatchException(String message) {
+        super(message);
+    }
+
+}
